+++ src/PhotoEffects.jl
@@ -53,8 +53,10 @@ include("effects/dither.jl")
 include("effects/contour.jl")
 include("effects/hatching.jl")
 include("effects/lineart.jl")
+include("effects/blobs.jl")
 
-export AbstractEffect, Brushes, Contour, Cubist, Dither, Duotone, Halftone,
+export AbstractEffect, Blobs, Brushes, Contour, Cubist, Dither, Duotone,
+       Halftone,
        Hatching,
        HexMosaic,
        LineArt, LowPoly,
@@ -84,6 +86,7 @@ export Seeding, Scatter, Given, sow, render, frame
         apply(Contour(), img)
         apply(Hatching(), img)
         apply(LineArt(), img)
+        apply(Blobs(colors = 2, blobs = 2, radius = 8), img)
         apply(Pipeline(Posterize(levels = 4), Duotone()), img)
         apply(VoronoiStained(points = 20), img)
         apply(VoronoiLloyd(points = 20, iterations = 1), img)